requirements.txt
kaggle
pandas
pyarrow
jupyter
python-dotenv
matplotlib
seaborn
scikit-learn
streamlit
